+++ CHANGELOG.rst
@@ -5,6 +5,17 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
+0.6.6
+*****
+* Made Auto-FOX compatible with Python 3.6.
+* Added tests for Python 3.6 and 3.8.
+* Permanently moved a number of modules from (nano-)CAT to Auto-FOX.
+* Added the ``MutliMolecule.loc`` property; allowing for the slicing of
+  MultiMolecule (directly) using atomic symbols.
+  Usage examples: ``mol.loc['Cd']`` and ``mol.loc['Cd', 'Se', 'O']``.
+  The Equivalent to ``mol[mol.atoms['Cd']]``.
+
+
 0.6.5
 *****
 * Cleaned up the angular distribution code & atom subset code.